fix : 수익률 출력 수정
수익률이 무조건 0으로 나오던 문제 수정
연산 이전에 double형으로 형변환 하도록 수정

+++ src/main/java/lotto/MatchResult.java
@@ -63,6 +63,6 @@ public class MatchResult {
         for (int i = 0; i < 5; i++) {
             rewards += prizeCount.get(i) * PRIZE_REWARDS[i];
         }
-        return rewards / (tryCount * 1000);
+        return (double) rewards / (tryCount * 1000);
     }
 }
